+++ src/main/java/kitchenpos/order/domain/repository/OrderRepository.java
@@ -2,7 +2,7 @@ package kitchenpos.order.domain.repository;
 
 import java.util.List;
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
