+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
 import java.util.Objects;
 import javax.persistence.Column;
