+++ src/main/java/kitchenpos/table/dto/request/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dto.request;
+package kitchenpos.table.dto.request;
 
 import java.util.List;
