+++ src/main/java/kitchenpos/table/dto/response/OrderTableResponse.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.dto.response;
+package kitchenpos.table.dto.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public class OrderTableResponse {
