+++ src/test/java/kitchenpos/acceptance/order/OrderAcceptanceTest.java
@@ -11,10 +11,10 @@ import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.Product;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
 import kitchenpos.order.dto.request.OrderLineItemRequest;
 import kitchenpos.order.dto.request.OrderRequest;
 import kitchenpos.order.dto.response.OrderResponse;
+import kitchenpos.table.domain.OrderTable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
