+++ src/test/java/kitchenpos/acceptance/table/TableGroupAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance.order;
+package kitchenpos.acceptance.table;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,11 +6,11 @@ import java.util.Arrays;
 import kitchenpos.acceptance.AcceptanceTest;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableGroup;
-import kitchenpos.order.dto.request.OrderTableRequest;
-import kitchenpos.order.dto.request.TableGroupRequest;
-import kitchenpos.order.dto.response.TableGroupResponse;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.dto.request.OrderTableRequest;
+import kitchenpos.table.dto.request.TableGroupRequest;
+import kitchenpos.table.dto.response.TableGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpEntity;
