+++ src/test/java/kitchenpos/domain/table/OrderTableTest.java
@@ -1,10 +1,10 @@
-package kitchenpos.domain.order;
+package kitchenpos.domain.table;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
