+++ src/test/java/kitchenpos/domain/table/TablesTest.java
@@ -1,12 +1,12 @@
-package kitchenpos.domain.order;
+package kitchenpos.domain.table;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Arrays;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.Tables;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.Tables;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
